View all user info is up and running, made all 3 parts into individual components making it more modular. Making it modular fixed the bugs.
