+++ src/app/project/view-users-projects/view-users-projects.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ViewUsersProjectsComponent } from './view-users-projects.component';
+
+describe('ViewUsersProjectsComponent', () => {
+  let component: ViewUsersProjectsComponent;
+  let fixture: ComponentFixture<ViewUsersProjectsComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ViewUsersProjectsComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ViewUsersProjectsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
