+++ src/app/project/view-users/view-users.component.html
@@ -1,0 +1,47 @@
+<!-- View all users table, only open to admins -->
+<div id="view-project-layout" *ngIf="usersPage">
+  <!-- a field for admins to enter in a string to filter the table -->
+  <mat-form-field>
+    <input matInput (keyup)="applyUserFilter($event.target.value)" placeholder="Filter">
+  </mat-form-field>
+
+  <div class="table-container mat-elevation-z8 my-theme" color="primary">
+    <table mat-table [dataSource]="dataSourceUsers" multiTemplateDataRows matSort>
+
+      <!-- Username Column -->
+      <ng-container matColumnDef="username">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header> Username </th>
+        <td mat-cell *matCellDef="let user"> {{user.username}} </td>
+      </ng-container>
+
+      <!-- Firstname Column -->
+      <ng-container matColumnDef="firstName">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header> Firstname </th>
+        <td mat-cell *matCellDef="let user"> {{user.firstName}} </td>
+      </ng-container>
+
+      <!-- Lastname Column -->
+      <ng-container matColumnDef="lastName">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header> LastName </th>
+        <td mat-cell *matCellDef="let user"> {{user.lastName}} </td>
+      </ng-container>
+
+      <!-- Email Column -->
+      <ng-container matColumnDef="email">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header> Email </th>
+        <td mat-cell *matCellDef="let user"> {{user.email}} </td>
+      </ng-container>
+
+      <!-- Role Column -->
+      <ng-container matColumnDef="role">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header> Role </th>
+        <td mat-cell *matCellDef="let user"> {{user.role}} </td>
+      </ng-container>
+
+      <!-- Need mat-header-row/mat-row to define the columns in a mat table -->
+      <tr mat-header-row *matHeaderRowDef="displayedUserColumns; sticky: true"></tr>
+      <tr mat-row *matRowDef="let project; columns: displayedUserColumns;"></tr>
+    </table>
+    <mat-paginator [pageSize]="5" [pageSizeOptions]="[5, 10, 20]" showFirstLastButtons></mat-paginator>
+  </div>
+</div>
